+++ index.js
@@ -136,6 +136,26 @@ module.exports = function(source) {
 
   hb.JavaScriptCompiler = MyJavaScriptCompiler;
 
+  // Define custom visitor for further template AST parsing
+  var Visitor = handlebars.Visitor;
+  function InternalBlocksVisitor() {
+    this.partialBlocks = [];
+    this.inlineBlocks = [];
+  }
+
+  InternalBlocksVisitor.prototype = new Visitor();
+  InternalBlocksVisitor.prototype.PartialBlockStatement = function(partial) {
+    this.partialBlocks.push(partial.name.original);
+    Visitor.prototype.PartialBlockStatement.call(this, partial);
+  };
+  InternalBlocksVisitor.prototype.DecoratorBlock = function(partial) {
+    if (partial.path.original === 'inline') {
+      this.inlineBlocks.push(partial.params[0].value);
+    }
+
+    Visitor.prototype.DecoratorBlock.call(this, partial);
+  };
+
   // This is an async loader
   var loaderAsyncCallback = this.async();
 
@@ -167,16 +187,23 @@ module.exports = function(source) {
     // Precompile template
     var template = '';
 
+    // AST holder for current template
+    var ast = null;
+
+    // Compile options
+    var opts = assign({
+      knownHelpersOnly: !firstCompile,
+      // TODO: Remove these in next major release
+      preventIndent: !!query.preventIndent,
+      compat: !!query.compat
+    }, precompileOptions, {
+      knownHelpers: knownHelpers,
+    });
+
     try {
       if (source) {
-        template = hb.precompile(source, assign({
-          knownHelpersOnly: !firstCompile,
-          // TODO: Remove these in next major release
-          preventIndent: !!query.preventIndent,
-          compat: !!query.compat
-        }, precompileOptions, {
-          knownHelpers: knownHelpers,
-        }));
+        ast = hb.parse(source, opts);
+        template = hb.precompile(ast, opts);
       }
     } catch (err) {
       return loaderAsyncCallback(err);
@@ -286,7 +313,19 @@ module.exports = function(source) {
       } else {
         partialResolver(request, function(err, resolved){
           if(err) {
-            return partialCallback(err);
+            var visitor = new InternalBlocksVisitor();
+
+            visitor.accept(ast);
+
+            if (
+              visitor.inlineBlocks.indexOf(request) !== -1 ||
+              visitor.partialBlocks.indexOf(request) !== -1
+            ) {
+              return partialCallback();
+            } else {
+              return partialCallback(err);
+            }
+
           }
           foundPartials[partial] = resolved;
           needRecompile = true;